Refactor main.dart and add MyApp class
- Refactor main.dart to remove MyApp class and move it to a separate file called my_app.dart
- Update imports in main.dart to import my_app.dart
- Update main() function in main.dart to use MyApp from my_app.dart
